Extended 'format' parameter with 'raw', to send unadulterated data.
This was necessary to address a POC issue. We're transmitting messages in
a format specified by a hardware vender. The vender cannot interpret
messages sent in the 'line' or 'block' format.

Extended the format parameter with 'raw', it indicates that the data
should be transmitted unaugmented.

Added a sample with explaination of what to look when using.

	modified:   TCPServer.xml
	modified:   TCPServer_cpp.cgt
	modified:   TCPServer_h.cgt
	new file:   ../../samples/TCPServerSendRaw/sample/TCPServerSendRaw.spl